chore(signage-manager): move tests to separate folder
