+++ apps/signage-manager/src/tests/shared/nav-items.spec.ts
@@ -1,9 +1,11 @@
-import { filterManageNavItems } from './nav-items';
+import { filterManageNavItems } from '../../app/shared/nav-items';
 
 describe('filterManageNavItems', () => {
     it('hides group management when the user cannot manage groups', () => {
         expect(
-            filterManageNavItems(false).some((item) => item.route === '/groups'),
+            filterManageNavItems(false).some(
+                (item) => item.route === '/groups',
+            ),
         ).toBe(false);
     });
